--- Assets/Scripts/GameSystem/WorldObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 135d384526760a54fbc12b2e99f93b75
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
